Our GitHub Action should ignore diff/*.pdf errors
Due to unchanged nearby tables and other reasons, latexdiff will sometimes
spuriously fail. CI should succeed nonetheless and just omit the diff PDF.
This is why we use `make -k`, but it does not affect make's exit status.

Extend run.sh so that it can ignore exit statuses. We could have just
used `make -i` or arranged to use `-latexdiff-vc` in the recipe, but doing
this in run.sh allows us to separate the two `make` commands so they are
tidily interleaved with their output in the logs.

(Somewhat unfortunately because action arguments can't be lists of strings
this increases the coupling between pr-pdfs.yaml and action.yaml.)